+++ pyannote/audio/models/separation/__init__.py
@@ -1,6 +1,6 @@
 # MIT License
 #
-# Copyright (c) 2020-2022 CNRS
+# Copyright (c) 2024- CNRS
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -20,18 +20,6 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from .multilabel import MultiLabelSegmentation
-from .overlapped_speech_detection import OverlappedSpeechDetection
-from .resegmentation import Resegmentation
-from .speaker_diarization import SpeakerDiarization
-from .speech_separation import SpeechSeparation
-from .voice_activity_detection import VoiceActivityDetection
+from .ToTaToNet import ToTaToNet
 
-__all__ = [
-    "VoiceActivityDetection",
-    "OverlappedSpeechDetection",
-    "SpeakerDiarization",
-    "Resegmentation",
-    "MultiLabelSegmentation",
-    "SpeechSeparation",
-]
+__all__ = ["ToTaToNet"]